Move `AreaTypesEnum` to `noris.types` and update imports accordingly.
