+++ nest-tax-backend/src/noris/utils/__tests__/tax-detail.helper.spec.ts
@@ -1,5 +1,6 @@
 import { NorisTaxPayersDto } from '../../noris.dto'
-import { AreaTypesEnum, mapNorisToTaxDetailData } from '../noris.helper'
+import { mapNorisToTaxDetailData } from '../noris.helper'
+import {AreaTypesEnum} from "../../noris.types";
 
 describe('taxDetail', () => {
   const mockTaxId = 123
